chore: move client transfer traits to own files
- This helps with code readability.
