+++ safenode/src/client/wallet/verifying_client.rs
@@ -10,105 +10,14 @@ use super::Client;
 
 use crate::{
     domain::{
-        client_transfers::{
-            create_online_transfer, Outputs as TransferDetails, SpendRequestParams,
-        },
         storage::dbc_address,
-        wallet::{Error, Result, SendClient, SendWallet, VerifyingClient},
+        wallet::{Error, Result, VerifyingClient},
     },
     network::close_group_majority,
-    protocol::messages::{Cmd, CmdResponse, Query, QueryResponse, Request, Response, SpendQuery},
+    protocol::messages::{Query, QueryResponse, Request, Response, SpendQuery},
 };
 
-use sn_dbc::{Dbc, DbcIdSource, DerivedKey, PublicAddress, Token};
-
-/// A wallet client can be used to send and
-/// receive tokens to/from other wallets.
-pub struct WalletClient<W: SendWallet> {
-    client: Client,
-    wallet: W,
-}
-
-impl<W: SendWallet> WalletClient<W> {
-    /// Create a new wallet client.
-    pub fn new(client: Client, wallet: W) -> Self {
-        Self { client, wallet }
-    }
-
-    /// Send tokens to another wallet.
-    pub async fn send(&mut self, amount: Token, to: PublicAddress) -> Result<Dbc> {
-        let dbcs = self.wallet.send(vec![(amount, to)], &self.client).await?;
-        match &dbcs[..] {
-            [info, ..] => Ok(info.dbc.clone()),
-            [] => Err(Error::CouldNotSendTokens(
-                "No DBCs were returned from the wallet.".into(),
-            )),
-        }
-    }
-
-    /// Return the wallet.
-    pub fn into_wallet(self) -> W {
-        self.wallet
-    }
-}
-
-#[async_trait::async_trait]
-impl SendClient for Client {
-    async fn send(
-        &self,
-        dbcs: Vec<(Dbc, DerivedKey)>,
-        to: Vec<(Token, DbcIdSource)>,
-        change_to: PublicAddress,
-    ) -> Result<TransferDetails> {
-        let transfer = create_online_transfer(dbcs, to, change_to, self).await?;
-
-        for spend_request_params in transfer.all_spend_request_params.clone() {
-            let SpendRequestParams {
-                signed_spend,
-                parent_tx,
-                fee_ciphers,
-            } = spend_request_params;
-
-            let cmd = Cmd::SpendDbc {
-                signed_spend: Box::new(signed_spend),
-                parent_tx: Box::new(parent_tx),
-                fee_ciphers,
-            };
-
-            let responses = self
-                .send_to_closest(Request::Cmd(cmd))
-                .await
-                .map_err(|err| Error::CouldNotSendTokens(err.to_string()))?;
-
-            // Get all Ok results of the expected response type `Spend`.
-            let ok_responses: Vec<_> = responses
-                .iter()
-                .flatten()
-                .flat_map(|resp| {
-                    if let Response::Cmd(CmdResponse::Spend(Ok(()))) = resp {
-                        Some(())
-                    } else {
-                        println!("Spend error {resp:?}.");
-                        None
-                    }
-                })
-                .collect();
-
-            // We require a majority of the close group to respond with Ok.
-            if ok_responses.len() >= close_group_majority() {
-                continue;
-            } else {
-                return Err(Error::CouldNotVerifyTransfer(format!(
-                    "Not enough close group nodes accepted the spend. Got {}, required: {}.",
-                    ok_responses.len(),
-                    close_group_majority()
-                )));
-            }
-        }
-
-        Ok(transfer)
-    }
-}
+use sn_dbc::Dbc;
 
 #[async_trait::async_trait]
 impl VerifyingClient for Client {
